+++ source/Octopus.Tentacle/Configuration/Instances/ApplicationInstanceSelector.cs
@@ -1,15 +1,9 @@
 using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Octopus.Configuration;
 using Octopus.Diagnostics;
 using Octopus.Tentacle.Util;
-#if !NETFRAMEWORK
-using k8s;
-using k8s.Models;
-using Newtonsoft.Json;
-#endif
 
 namespace Octopus.Tentacle.Configuration.Instances
 {
@@ -164,88 +158,4 @@ namespace Octopus.Tentacle.Configuration.Instances
             }
         }
     }
-
-#if !NETFRAMEWORK
-    class ConfigMapKeyValueStore : IWritableKeyValueStore, IAggregatableKeyValueStore
-    {
-        const string Namespace = "octopus";
-        readonly string name;
-        readonly k8s.Kubernetes client;
-
-        V1ConfigMap configMap;
-
-        IDictionary<string, string> ConfigMapData => configMap.Data ?? (configMap.Data = new Dictionary<string, string>());
-
-        public ConfigMapKeyValueStore(string instanceName)
-        {
-            var kubeConfig = KubernetesClientConfiguration.InClusterConfig();
-            client = new k8s.Kubernetes(kubeConfig);
-            name = $"{instanceName.ToLowerInvariant()}-configmap";
-            V1ConfigMap? config;
-            try
-            {
-                config = client.CoreV1.ReadNamespacedConfigMap(name, Namespace);
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine($"ConfigMapKeyValueStore.ctor Exception: {e}");
-                config = null;
-            }
-
-            configMap = config ?? client.CoreV1.CreateNamespacedConfigMap(new V1ConfigMap { Metadata = new V1ObjectMeta { Name = name, NamespaceProperty = Namespace }}, Namespace);
-            Console.WriteLine($"ConfigMapKeyValueStore.ctor ConfigMap loaded: {JsonConvert.SerializeObject(configMap)}");
-        }
-
-        public string? Get(string name, ProtectionLevel protectionLevel = ProtectionLevel.None)
-        {
-            return ConfigMapData.TryGetValue(name, out var value) ? value : null;
-        }
-
-        public TData? Get<TData>(string name, TData? defaultValue = default, ProtectionLevel protectionLevel = ProtectionLevel.None)
-        {
-            var result = TryGet<TData>(name, protectionLevel);
-
-            return result.foundResult ? result.value : defaultValue;
-        }
-
-        public bool Set(string name, string? value, ProtectionLevel protectionLevel = ProtectionLevel.None)
-        {
-            configMap.Data[name] = value;
-            return Save();
-        }
-
-        public bool Set<TData>(string name, TData value, ProtectionLevel protectionLevel = ProtectionLevel.None)
-        {
-            return Set(name, JsonConvert.SerializeObject(value), protectionLevel);
-        }
-
-        public bool Remove(string name)
-        {
-            return ConfigMapData.Remove(name) && Save();
-        }
-
-        public bool Save()
-        {
-            configMap = client.CoreV1.ReplaceNamespacedConfigMap(configMap, name, Namespace);
-            return true;
-        }
-
-        public (bool foundResult, TData? value) TryGet<TData>(string name, ProtectionLevel protectionLevel = ProtectionLevel.None)
-        {
-            var value = Get(name, protectionLevel);
-
-            if (value == null) return (false, default);
-
-            try
-            {
-                return value is TData data ? (true, data) : (true, JsonConvert.DeserializeObject<TData>(value));
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine($"ConfigMapKeyValueStore.TryGet<TData>: Exception! {e}");
-                return (false, default);
-            }
-        }
-    }
-#endif
 }
